refactor: Renamed  utils/tests to utils/enzyme
